sv_parser.c: check packet size before copying
A buffer is used to store all the information of a packet after parsing.
This buffer is only allocated once, to reduce execution time. It's size
is determined by the maximum size of a packet in the IEC-61850 norm.

However, if a malformed too long packet is received, it could write
outside of the buffer and lead to remote code execution.
This commit fixes this issue by returning BAD_FORMAT in case of a too
long packet. It also uses the unlikely C compiler feature to reduce the
execution time overhead.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>